The table displays as a table, as was originally intended, and has also been improved further.
The table now actually displays as a table. Also, the text has been made better. Here's two examples: 1. A typo in which the instructions for Flash 4k Game Competition to go to Contest Rules on the web page being in the "Website link/links" column instead of the "Other rules" column, as well as another typo in which the "Amount of competitions" for the 5k isn't shown quite how I intended to do so, has been fixed. 2. Website links have been merged into other text than the website links. This makes the table less wide, reducing the likelihood of needing to scroll left and right to see more of the table, and look more neat and tidy too.