Some naming history for MiniGame Compo instead of one name
Instead of this project calling MiniGame Compo by the name MiniGame Compo, this project now calls MiniGame Compo by whatever names it was called throughout time. However, it only gives the names for the 2010-2013 versions of the competition due to time constraints, but I plan to eventually add the rest of the names of the competition.