2002-2009 Minigamecompo/Minigame Compo names added, js4kgames "Amount of competitions" description shortened
Every name for every Minigamecompo/Minigame Compo contest held from 2002-2009 has been added. Also, the js4kgames "Amount of competitions" description has been shortened from "None (however, one was announced for 2016, as both [this Tweet](http://twitter.com/js4kgames/status/651873094416601091) and [this tweet](http://twitter.com/js4kgames/status/651875841975128064) says (the former tweet of the two is more specific, saying May 2016))" to "None (however, one was announced for 2016, as both [this Tweet](http://twitter.com/js4kgames/status/651873094416601091) and [this tweet](http://twitter.com/js4kgames/status/651875841975128064) says, the former of which is more specific, saying May 2016)", with a typo in "as both [this tweet]" in which "tweet" was misspelled as "Tweet" being fixed as well, thus the new js4kgames "Amount of competitions" description is actually "None (however, one was announced for 2016, as both [this tweet](http://twitter.com/js4kgames/status/651873094416601091) and [this tweet](http://twitter.com/js4kgames/status/651875841975128064) says, the former of which is more specific, saying May 2016)". Please note the descriptions given here are in this README.md's markup language, which, as of the release of this commit change, is GFM (GitHub Flavored Markdown), a variant of CommonMark, a variant of Markdown that aims to standardize Markdown.